Update README.md
The standard ddev-webserver comes with PHP pgsql extension enabled, but lacks the psql client. This little fix to your config file will install the  extra webimage package for postgresql-11. This is especially important when you run Drush for a Drupal site on a PostgreSQL database, which will fail if the psql command isn't available.