Expand and clarify README.md. Correct typos.
Added ~/ to indicate the global .ddev folder level that was being affected.

Added ~/ to clarify and allow functional cut and paste of the command.

Added command to rename id_rsa.pub to authorized keys, to clarify and complete the necesscary work, and allow one cut and paste start and finish the work.

Clarified for user that the project .ddev/homeadditions/ folder could be used as well, because it was a stumbling block for me and knowledge gained along the way. At first I was manually performing the mkdir, cp, mv operation at the project level and it worked but I had to look closer to see the command was written for the global(user?) level.