Updated Ports - Changed from ports to expose
Changed 'ports' to 'expose' in docker-composer.chrome.yaml. Ports exposed to the other services but not to the host machine.

That approach usually isn’t sustainable because two projects might want to use the same port, so we expose the additional port to the Docker network and then use ddev-router to bind it to the host. This works only for services with an HTTP API, but results in having both HTTP and HTTPS ports (9998 and 9999).

https://ddev.readthedocs.io/en/latest/users/extend/custom-compose-files/#docker-composeyaml-examples